// backend/src/common/enums/booking-status.enum.ts
export enum BookingStatus {
  PENDING = 'PENDING',
  AWAITING_GUIDE_ASSIGNMENT = 'AWAITING_GUIDE_ASSIGNMENT',
  GUIDE_ASSIGNED = 'GUIDE_ASSIGNED',
  GUIDE_ACCEPTED = 'GUIDE_ACCEPTED',
  GUIDE_REJECTED = 'GUIDE_REJECTED',
  CONFIRMED = 'CONFIRMED',
  ONGOING = 'ONGOING',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}